# improve yaml parsing performance by ~25% by using PyYAML's CParser loader