This is with PyYAML-3.11, the CParser loader is a good bit faster than the python loader in my tests.

With SafeParser:
`% time ansible-playbook -l dummyNonConnectingHost site.yml --vault-password-file=~/.vault_pass > /dev/null
613.76s user 86.28s system 97% cpu 11:54.65 total`

With CSafeParser:
`% time ansible-playbook -l dummyNonConnectingHost site.yml --vault-password-file=~/.vault_pass > /dev/null
439.21s user 82.81s system 98% cpu 8:52.45 total`