@abadger 2015-02-16
Thanks!   I think we had assumed that the yaml library was using the C extension if that was available but your benchmarks and a review of the code show that we have to be explicit about using it.